padoa-tools/argocd-tools/requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
Deprecated==1.2.10
flatten-dict==0.3.0
humanfriendly==8.2
idna==2.10
pathlib2==2.3.5
PyGithub==1.51
PyJWT==1.7.1
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
spinners==0.0.24
urllib3==1.26.6
wrapt==1.12.1
yaspin==0.18.0
asyncio
websockets
questionary
azure-storage-blob
redis==3.5.3
click==8.0.1
